requirements.txt
boto3==1.24.81
botocore==1.27.81
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
faiss-cpu==1.7.2
filelock==3.8.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
s3transfer==0.6.0
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
tokenizers==0.5.2
torch==1.12.1
tqdm==4.64.1
transformers==2.5.1
typing_extensions==4.3.0
urllib3==1.26.12
